Match Linux x86 architecture name mapping and add Debian mapping (#1)
Previously, the Linux naming convention translated the `x86_64`
architecture to `amd64`, resulting in the following runtime triple:

   amd64-unknown-linux-gnu

However Swift does not recognise this triple:

   error: emit-module command failed with exit code 1 (use -v to see invocation)
   <unknown>:0: error: could not find module '_Concurrency' for target 'amd64-unknown-linux-gnu'; found: x86_64-unknown-linux-gnu, at: ...

Almost all Swift triples call this architecture 'x86_64' - the
exception is OpenBSD, which uses 'amd64'.

In fact, Swift on macOS, Swift on Linux and Debian packages all use
different conventions to refer to the x86 and Arm architectures:

                   |  x86_64    arm64
  ------------------------------------
  Swift macOS      |  x86_64    arm64
  Swift Linux      |  x86_64  aarch64
  Debian packages  |   amd64    arm64

Mapping `x86_64` to `amd64` is necessary when downloading Debian
packages, but not when constructing a Linux cross-compilation SDK
bundle.   Conversely, mapping `arm64` to `aarch64` is necessary
when constucting the SDK bundle but not when downloading Debian
packages.

This commit changes the `linuxConventionName` mapping to match the
Swift SDK mapping and adds a `debianConventionName` mapping which is
used when downloading packages.