Remove protocol indirection for `SwiftSDKGenerator`
There is only a single implementation this protocol: `LocalSwiftSDKGenerator`. The original intent was to introduce additional implementations for testing, but I no longer think that this protocol on its own is useful. We should protocolize on a different level, i.e. file system access, process spawning etc. That will be realized in subsequent PRs.