Use standard PATH to search for docker and other commands
SwiftSDKGenerator expects the `docker` CLI tool to have been installed
using Homebrew.   It also assumes that Homebrew is installed in the
default location for an x86_64 or arm64 host.   This will not work
if the user has installed Homebrew in a custom location, or is using
the `docker` CLI installed by Docker Desktop.  Using the PATH
inherited from the environment avoids these problems and allows the
user to control which tools will be used.