Add `GeneratorEngine` module with tests and macros (#31)
We'd like to avoid redundant rebuilding and regeneration for inputs that don't change. The new `GeneratorEngine` module achieves that by introducing a `Query` abstraction, which can be hashed and matched to its cached outputs if any are available.

The new `Engine` actor can run such `Query` and immediately return a cached result file if one is present and its hash matches the recorded one.

This required introducing new `@Query` and `@CacheKey` macros for making hashing easy and consistent on `QueryProtocol` types and their inputs.  We can't use the standard `Hashable` protocol, as that produces new hashes on every process launch due to randomized hashing seeds. With Swift Crypto's `HashFunction` we can generate consistent hashes and record them in `SQLiteBackedCache`.

To make the caching engine testable, I've introduced a new `FileSystem` actor protocol with `VirtualFileSystem` and `LocalFileSystem` actor implementations.