Cache `lld` binaries to avoid redundant rebuilds (#35)
This proceeds with converting more of the generation steps to a query-based architecture. Now `lld` doesn't need to be rebuilt if its source directory and build options don't change. This is implemented with a new `CMakeBuildQuery`, which takes source directory, and build options as arguments, and returns a freshly-built binary that can be cached by the generator engine.

Also renamed `DownloadQuery` to `DownloadArtifactQuery` and moved it to a separate file.