docker: Copy architecture-specific library directories from Ubuntu images
Ubuntu's multiarch directory scheme puts many libraries in
architecture-specific directories:

   https://wiki.ubuntu.com/MultiarchSpec

Without these directories the linker cannot find several critical libraries,
including the C library and initial runtime support:

    % swift build --experimental-swift-sdk 5.9-RELEASE_ubuntu_jammy_x86_64
    Building for debugging...
    ...
    error: link command failed with exit code 1 (use -v to see invocation)
    ld.lld: error: cannot open Scrt1.o: No such file or directory
    ld.lld: error: cannot open crti.o: No such file or directory
    ld.lld: error: unable to find library -lc
    ld.lld: error: cannot open crtn.o: No such file or directory
    clang-13: error: linker command failed with exit code 1 (use -v to see invocation)
    error: fatalError
    [5/6] Linking hello-world

This change allows the linker to find the C library but it is still
unable to find ld-linux.so.   That will be the subject of the next
commit.