Cancel tasks when receiving `SIGINT` signal
This adds a dependency on the `UnixSignal` product from https://github.com/swift-server/swift-service-lifecycle.

Additionally, had to slightly restructure `GeneratorCLI` to explicitly track possible state in the new `enum GeneratorCLI.State`. With that we can create a stream of state changes and launch two unstructured tasks concurrently:

1. The generator task.
2. The signals sequence reader task.

The reason for keeping these tasks unstructured is that we need to hold handles to these for explicit cancellation. Then if the signal is produced by `UnixSignalSequence` first, we can cancel the generator task. If the generator finishes successfully, we cancel the signals reader task and shutdown normally. If the generator fails with an error, we rethrow it from the task to the top level.