generator: Disable all libllvm targets when building lld
lld is always built as a full cross linker, able to handle all
architectures and binary formats:

> It is always a cross-linker, meaning that it always supports all
> the above targets however it was built. In fact, we don’t provide
> a build-time option to enable/disable each target.

-- https://lld.llvm.org/#

It does not seem to be necessary to build _any_ LLVM targets.   With
LLVM_TARGETS_TO_BUILD set to the empty string, a viable lld binary
is still built.   On my machine this speeds up a cold-cache build by
another 4 minutes:

* Before: ninja builds 1682 edges; overall SDK build is about 18
  minutes.
* After: ninja builds 1492 edges; overall SDK build is about 14
minutes.