generator: Don't overwrite libc.so linker script on RHEL
A Swift program which called `atexit` failed to link using a RHEL
SDK, but the same program would link successfully using an Ubuntu SDK.

    ld.lld: error: undefined symbol: atexit

`atexit` comes from the C library but is not part of the
dynamically-linked `libc.so.6`.  Instead it is defined in a
statically-linked library `libc_nonshared.a`.    This library
is present in the RHEL SDK but was not being found by `lld`.

This was happening because - only for RHEL SDKs - the generator was
overwriting `libc.so` with a symlink to `libc.so.6`.  `libc.so` is
in fact a GNU ld script, not a library.   When the linker searches
for `-lc`, it finds the script which tells it to use the real library
at `libc.so.6`.   Crucially it _also_ tells the linker to look in
`libc_nonshared.a` if the symbols it needs are not in `libc.so.6`:

    /* GNU ld script
       Use the shared library, but some functions are only in
       the static library, so try that secondarily.  */
    OUTPUT_FORMAT(elf64-x86-64)
    GROUP ( /lib64/libc.so.6 /usr/lib64/libc_nonshared.a  AS_NEEDED ( /lib64/ld-linux-x86-64.so.2 ) )

The generator was replacing the `libc.so` linker script because it
contained absolute paths.   However it turns out that `libc.so` in
the Ubuntu SDK is also a script with similar absolute paths:

    /* GNU ld script
       Use the shared library, but some functions are only in
       the static library, so try that secondarily.  */
    OUTPUT_FORMAT(elf64-x86-64)
    GROUP ( /lib/x86_64-linux-gnu/libc.so.6 /usr/lib/x86_64-linux-gnu/libc_nonshared.a  AS_NEEDED ( /lib64/ld-linux-x86-64.so.2 ) )

These absolute paths seem not to cause problems because the linker
is called with the `--sysroot` argument pointing to the root of the
SDK, which changes `lld`'s search path.   This can be seen by running
`swift build -v` to print the (enormous) linker command line;  running
the command again without the `--sysroot` flag causes the linker
to search for libraries in the host root directory and fail.

It turns out that `libm.so` is also a linker script with similar
contents to `libc.so`.   A couple of other libraries also have linker
scripts.

This commit extends the EndToEnd tests to iterate over a list of
test programs, and adds a test to check that `atexit` can be linked.