tests: Build on Linux even though tests are not currently run
The 'EndToEndTests' are only run on Linux but testing is currently
disabled in CI.  This makes it likely that the tests will not be
built often and makes it possible for them to drift out of sync with
the code they test without being noticed.

This change ensures that the tests are built in CI, even if they
are not run.