Make `supportedTriples` optional in SDK metadata for universal SDKs
For SDKs that can be universally used on any host platform, the
`supportedTriples` field in the SDK metadata was previously required to
fill all possible host triples.
This change makes the `supportedTriples` field optional, allowing SDKs
to be universally used without specifying any host triples, and Wasm SDK
generator fills the field with null when `--host-swift-package-path` is
not provided.