tests: Run tests on Linux CI (#146)
* QueryEngine: Calculate stable hashes for Foundation.URL and FoundationEssentials.URL

Until Swift 6.0, the fully-qualified name of the URL type was
`Foundation.URL`.  After the adoption of FoundationEssentials, the
name changed to `FoundationEssentials.URL`.  This difference causes
the hashes to change, so for backwards compatibility we pin the
type name to `Foundation.URL`.

Fixes #45

* tests: Run tests on Linux CI

The EndToEnd tests cannot currently be run:

* they deadlock when run under `swift test` (#143)
* the CI requires the use of an HTTP proxy to download from the
  Internet, but SDK generator does not support proxies (#145)

The tests are not compiled on macOS, but this commit marks
them with `XCTSkip` so they are not run on Linux either.

Future commits will loosen these restrictions.

Fixes #45