tests: Fix EndToEnd test deadlock under `swift test`, but exclude from CI (#144)
Building an SDK requires running the sdk-generator with `swift run swift-sdk-generator`.  This takes a lock on `.build`, but if the tests are being run by `swift test` the outer Swift Package Manager instance will already hold this lock, causing the test to deadlock. We can avoid this by giving the `swift run swift-sdk-generator` instance its own scratch directory.

This PR consolidates the code which builds SDKs in one function, then runs each SDK build with its own temporary scratch directory.   This avoids the deadlock on the `.build` directory when run under `swift test`.

The tests continue to pass when run under Xcode, which was not affected by the deadlock because it works in a separate scratch directory.

:warning: **The tests cannot currently run in CI because SDK generator cannot use the HTTP proxy to download packages.**   This PR allows the tests to be run locally with `swift test` but skips them when running in the CI.   When the downloading problem is solved, the tests can be enabled in CI as well.

:warning: RHEL-based Swift 6.0 SDKs built from container images currently do not currently work, as reported in #138.   This PR makes it possible to run the test locally under `swift test` while working on this problem, to prevent regressions such as #141.

Fixes #143