generator/linux: Use `--ldpath` linker flag for 5.9 (#151)
Swift Package Manager's linker flag handling changed between 5.9 and 5.10.   The flags which work for 5.9 cause linking failures 5.10 and later, and vice versa:

  https://github.com/swiftlang/swift-package-manager/issues/7222

This commit generates workaround flags for 5.9 and new-style flags for all other versions.

The EndToEnd tests currently cannot run in CI, so this change was tested locally. (Issue #145)

* The basic 'hello world' example generated by `swift package init` built successfully with 5.9.2, 5.10.1 and 6.0.2 SDKs on x86_64 and aarch64.
* A more complex example using Vapor built succesfully with 5.9.2, 5.10.1.  6.0.2 failed because of the CShims problem reported in Issue #138.