generator: Do not include the host toolchain by default
Not including the host toolchain matches the Swift Static Linux
SDK.  Users with experience of the static SDK will be familiar with
the need to install a matching OSS toolchain to use it:

   https://www.swift.org/documentation/articles/static-linux-getting-started.html#installing-the-sdk

Not including the host toolchain reduces the size of the SDK
bundle by about 2GB.   This makes SDKs easier to move around,
wastes less storage space if multiple SDKs are installed, and
speeds up the SDK build process - especially the first build,
as the host toolchain does not have to be downloaded.

WASM SDKs also do not include the host toolchain.