generator: --from-container-image implies --with-docker (#174)
If --from-container-image is specified without --with-docker,
the generator builds an SDK from Debian packages, completely
ignoring the user's clear intent to build from a container
image.

For now, we can assume that a user who specifies --from-container-image
also wants to use a container runtime.  In future we might be able
to support extracting container images without the need for a
runtime.   If that happens, we can stop assuming --with-docker
and commands which only specify --from-container-image will continue
to work, so scripts and documentation won't need to be changed.