generator: Do not attempt to download toolchain if it is not needed
SDKs now do not include the host toolchain by default, but the generator
still checks for host toolchains before deciding it does not need to
download them.

Skipping this check when a host toolchain is not requested removes the
unnecessary  network request.    This saves a couple of seconds but also
makes it possible to run the generator offline, if building from a container
image which is already available locally.