Fix support for building with `--static-swift-stdlib` for armv7 (#194)
The [atomic library is used in the Swift stdlib](https://github.com/swiftlang/swift/blob/69756635819114a851023aa2e8e6fd87f51d094d/cmake/modules/AddSwift.cmake#L336-L338) for 32-bit architectures. As such, when attempting to compile a binary containing the static Swift stdlib using a Swift SDK for armv7, `-latomic` must be linked. This is not needed for regular cross-compilation, just when passing `--static-swift-stdlib`.

I've added `-latomic` for "armv7" and added a test case in `LinuxRecipeTests` to ensure that it is included in the toolset.json file.

Now, I can build my binary that contains the statically linked Swift stdlib with this command:

```
swift build --swift-sdk 6.0.3-RELEASE_debian_bookworm_armv7 --static-swift-stdlib
```

This generates a rather large binary when also linking other things such as Foundation and Dispatch, as is to be expected:

```
$ pwd
~/tmp/hummingbird-examples/hello
$ du -hs .build/debug/App
123M	.build/debug/App
```

But that's fine, because it works perfectly on the target (Raspberry Pi 2) and provides yet another option for compiling binaries using the Swift SDK.