workflows: Run unit tests using GitHub Actions (#200)
* workflows: Re-enable tests job

* workflows: Disable tests on Windows

Requires NIO support on Windows

* workflows: Disable end-to-end tests in GitHub Actions

There is a good chance that GitHub Actions will be able to run these tests,
but we will disable them for now in order to get the unit tests running.
The end-to-end tests could not run with swift-ci, so disabling them here
does not reduce CI coverage.