Add missing macOS dependencies to Brewfile, update README.md (#231)
I recently did a clean install of macOS on an Intel MacBook and tried to run the SDK generator on it. I found that a few dependencies were missing from the `Brewfile`, especially when trying to create a Swift SDK for 5.10.1, which requires LLVM to be compiled for the host to include in the SDK.

I've added these dependencies to the Brewfile and also updated the README.md to show what each one is for.

---------

Co-authored-by: Max Desiatov <m_desiatov@apple.com>