Always use `rawValue` when using OS enum as string
Certain enum cases, such as `freeBSD`, contain capital letters. However, target triples must be entirely in lower case. The enum case's raw value reflects this, but our custom initializer for the Triple structure does not use it, resulting in an invalid target triple. This makes the generated SDK unusable if the user passed only the desired architecture instead of a complete triple.

To address this problem, the initializer should always request the rawValue explicitly.